# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Carve API
  description: |
    The Carve language-learning API. This document is the source of truth for
    every `/v1` endpoint; the TypeScript client in `apps/web/src/lib/api.ts`
    is regenerated from it and schemathesis runs property-based HTTP fuzz
    against it as part of the L4 testing layer.
  version: 0.3.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.carve.app
    description: Production
security: []  # default: no auth; protected operations explicitly set `security: [{ bearer: [] }]`
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Ok:
      type: object
      properties:
        ok: { type: boolean }
...
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email, minLength: 3, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 100 }
    AuthRegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email:        { type: string, format: email }
        password:     { type: string, minLength: 8, maxLength: 100 }
        display_name: { type: string, minLength: 1, maxLength: 80 }
    AuthTokenResponse:
      type: object
      required: [access_token, expires_in, user]
      properties:
        access_token: { type: string }
        expires_in:   { type: integer, description: "Lifetime of access_token in seconds" }
        user:         { $ref: '#/components/schemas/AuthUser' }
...
paths:
...
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/AuthLoginRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/AuthRegisterRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Email taken, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Refresh the access token using the refresh-token cookie
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenResponse' } } } }
        '401': { description: No / expired refresh token, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /v1/auth/forgot:
    post:
      tags: [auth]
      operationId: forgotPassword
      summary: Send a password-reset email
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ForgotPasswordRequest' } }
      responses:
        '200': { description: Always 200 to prevent email enumeration, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/reset:
    post:
      tags: [auth]
      operationId: resetPassword
      summary: Complete a password reset
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ResetPasswordRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
        '400': { description: Bad / expired token, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/verify:
    post:
      tags: [auth]
      operationId: verifyEmail
      summary: Mark the user's email as verified using a token
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/VerifyEmailRequest' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean }
        '400': { description: Bad / expired token, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/auth/verify/resend:
    post:
      tags: [auth]
      operationId: resendVerify
      summary: Resend the email-verification email
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ResendVerifyRequest' } }
      responses:
        '200':
          description: Always 200 to prevent enumeration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'

  # ── Users ─────────────────────────────────────────────────────────────
  /v1/users/me:
    get:
      tags: [users]
      operationId: getMe
      summary: Get the current user
      security: [{ bearer: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [users]
      operationId: updateMe
      summary: Update the current user
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/UserUpdateRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [users]
      operationId: deleteMe
      summary: Delete the current user (irreversible)
      security: [{ bearer: [] }]
      responses:
        '204': { description: Deleted }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Onboarding ────────────────────────────────────────────────────────
  /v1/onboarding/known-words:
    post:
      tags: [onboarding]
      operationId: onboardingKnownWords
      summary: Mark a batch of lemmas as known
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/OnboardingKnownWordsRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingKnownWordsResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/onboarding/starter-deck:
    post:
      tags: [onboarding]
      operationId: onboardingStarterDeck
      summary: Subscribe the user to the official starter deck for a language
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/OnboardingStarterDeckRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingStarterDeckResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Cards ─────────────────────────────────────────────────────────────
  /v1/cards:
    get:
      tags: [cards]
      operationId: listCards
      summary: List the user's cards
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: language
          schema: { $ref: '#/components/schemas/LanguageCode' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500 }
        - in: query
          name: state
          schema: { type: string, enum: [new, learning, review, relearning, suspended] }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CardListResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [cards]
      operationId: createCard
      summary: Create a card (mining)
      description: >
        Idempotent on (user, language, lemma). Returns 201 with the new card, or
        200 with the existing card (response includes "existing": true) when the
        word was already mined — so the client can attach media to it rather than
        creating a duplicate.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CardCreateRequest' } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CardCreateResponse' } } } }
        '200': { description: Existing card returned (already mined), content: { application/json: { schema: { $ref: '#/components/schemas/CardCreateResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/bulk:
    post:
      tags: [cards]
      operationId: bulkCards
      summary: Apply a bulk action to a set of cards
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CardBulkRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CardBulkResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    get:
      tags: [cards]
      operationId: getCard
      summary: Get a card by id
      security: [{ bearer: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Card' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [cards]
      operationId: updateCard
      summary: Update a card
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CardUpdateRequest' } }
      responses:
        '204': { description: Updated }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [cards]
      operationId: deleteCard
      summary: Delete a card
      security: [{ bearer: [] }]
      responses:
        '204': { description: Deleted }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}/media:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [cards]
      operationId: attachMedia
      summary: Upload image/audio (and subtitle metadata) for a card
      description: >
        Multipart upload. All parts are optional; the handler stores whatever is
        provided and returns the resulting URLs (null when not uploaded). If a
        part is supplied but the media backend fails to store it, the endpoint
        returns 502 (any part that did succeed is still persisted).
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:                { type: string, format: binary }
                audio:                { type: string, format: binary }
                subtitle_start_ms:    { type: integer }
                subtitle_end_ms:      { type: integer }
                video_source_url:     { type: string }
                subtitle_translation: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  image_url: { type: string, nullable: true }
                  audio_url: { type: string, nullable: true }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Card not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: A supplied media part failed to store, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}/suspend:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [cards]
      operationId: suspendCard
      summary: Suspend a card (skip in review until manually unsuspended)
      security: [{ bearer: [] }]
      responses:
        '200': { description: Suspended }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}/unsuspend:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [cards]
      operationId: unsuspendCard
      summary: Un-suspend a card
      security: [{ bearer: [] }]
      responses:
        '200': { description: Unsuspended }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}/bury:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [cards]
      operationId: buryCard
      summary: Bury a card until tomorrow
      security: [{ bearer: [] }]
      responses:
        '200': { description: Buried until tomorrow }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/cards/{id}/unbury:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [cards]
      operationId: unburyCard
      summary: Make a buried card immediately available again
      security: [{ bearer: [] }]
      responses:
        '200': { description: Unburied }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Review ────────────────────────────────────────────────────────────
  /v1/review/due-count:
    get:
      tags: [review]
      operationId: getDueCount
      summary: Count cards due now for the current user
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: language
          schema: { $ref: '#/components/schemas/LanguageCode' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DueCountResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/session:
    get:
      tags: [review]
      operationId: getReviewSession
      summary: Build the next review session (interleaved queue)
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: language
          schema: { $ref: '#/components/schemas/LanguageCode' }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewSessionResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/events:
    post:
      tags: [review]
      operationId: submitReviewEvent
      summary: Submit a review event
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ReviewEventRequest' } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewEventResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Card not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/undo:
    post:
      tags: [review]
      operationId: undoReview
      summary: Undo the user's most recent review event
      security: [{ bearer: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UndoResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Nothing to undo, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/intervals:
    get:
      tags: [review]
      operationId: getIntervals
      summary: Preview the next-due dates for each rating button
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: card_id
          required: true
          schema: { $ref: '#/components/schemas/UUID' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IntervalsResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Card not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/forecast:
    get:
      tags: [review]
      operationId: getForecast
      summary: 14-day workload forecast
      security: [{ bearer: [] }]
      parameters:
        - in: query
          name: language
          schema: { $ref: '#/components/schemas/LanguageCode' }
        - in: query
          name: days
          schema: { type: integer, minimum: 1, maximum: 90 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ForecastResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/notifications:
    get:
      tags: [review]
      operationId: listNotifications
      summary: List the user's pending review notifications (leeches, optimizer, etc.)
      security: [{ bearer: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationsResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/review/notifications/{id}/read:
    parameters:
      - in: path
        name: id
        required: true
        schema: { $ref: '#/components/schemas/UUID' }
    post:
      tags: [review]
      operationId: markNotificationRead
      summary: Mark a notification as read
      security: [{ bearer: [] }]
      responses:
        '204': { description: Marked read }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Decks ─────────────────────────────────────────────────────────────
...
  /v1/billing/subscription:
    get:
      tags: [billing]
      operationId: getSubscription
      summary: Get the user's current subscription
      security: [{ bearer: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Subscription' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...